feat: 리뷰 레포지토리에 어댑터 패턴 적용
